refactor: 순환참조 -> 퍼사드 패턴 사용
