--- src/main/java/com/sookmyung/concon/User/dto/UserRandomResponseDto.java
@@ -1,12 +1,0 @@
-package com.sookmyung.concon.User.dto;
-
-import com.sookmyung.concon.Order.entity.Orders;
-
-import java.util.List;
-import java.util.Optional;
-
-public class UserRandomResponseDto {
-    private Long userId;
-    private String username;
-    private List<Optional<Orders>> orders;
-}
